feat/backend: Refactored code to keep code DRY
