+++ backend/src/main/java/ca/bc/gov/app/core/entity/CodeDescr.java
@@ -1,0 +1,10 @@
+package ca.bc.gov.app.core.entity;
+
+import java.io.Serializable;
+
+public interface CodeDescr extends Serializable {
+
+	String getCode();
+	String getDescription();
+	
+}
