+++ backend/src/main/java/ca/bc/gov/app/core/util/impl/CoreUtilImpl.java
@@ -3,8 +3,10 @@ package ca.bc.gov.app.core.util.impl;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
+import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.apache.commons.collections.CollectionUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -15,7 +17,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 
 import ca.bc.gov.app.core.CoreConstant;
+import ca.bc.gov.app.core.entity.AbstractCodeDescrEntity;
 import ca.bc.gov.app.core.util.CoreUtil;
+import ca.bc.gov.app.core.vo.CodeDescrVO;
 
 @Component
 @Qualifier(CoreUtil.BEAN_NAME)
@@ -84,6 +88,26 @@ public class CoreUtilImpl implements CoreUtil {
 			}
 		}
 		return finalStr;
+	}
+
+	@Override
+	public List<CodeDescrVO> toSortedCodeDescrVOs(List<? extends AbstractCodeDescrEntity> codeDescrEntities) {
+		Comparator<AbstractCodeDescrEntity> comparatorByDescription = (e1, e2) ->
+				(e1.getDescription()).compareTo(e2.getDescription());
+		
+		return codeDescrEntities.
+				stream().
+				sorted(comparatorByDescription).
+				map(this::toCodeDescrVO).
+				collect(Collectors.toList());
+	}
+	
+	@Override
+	public CodeDescrVO toCodeDescrVO(AbstractCodeDescrEntity codeDescrEntity) {
+		return codeDescrEntity == null ? null : new CodeDescrVO(codeDescrEntity.getCode(), 
+																codeDescrEntity.getDescription(),
+																codeDescrEntity.getEffectiveDate(),
+																codeDescrEntity.getExpiryDate());
 	}
 	
 }
