+++ backend/src/main/java/ca/bc/gov/app/m/postgres/client/entity/SubmissionStatusCodeEntity.java
@@ -1,7 +1,5 @@
 package ca.bc.gov.app.m.postgres.client.entity;
 
-import java.util.Date;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -11,14 +9,14 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 import ca.bc.gov.app.core.configuration.PostgresPersistenceConfiguration;
-import ca.bc.gov.app.core.entity.AbstractEntity;
+import ca.bc.gov.app.core.entity.AbstractCodeDescrEntity;
 import ca.bc.gov.app.core.misc.scope.ScopeConstant;
 
 @Entity
 @Table(name = "SUBMISSION_STATUS_CODE", schema = PostgresPersistenceConfiguration.POSTGRES_ATTRIBUTE_SCHEMA)
 @Component(SubmissionStatusCodeEntity.BEAN_NAME)
 @Scope(ScopeConstant.PROTOTYPE)
-public class SubmissionStatusCodeEntity implements AbstractEntity {
+public class SubmissionStatusCodeEntity extends AbstractCodeDescrEntity {
 
 	private static final long serialVersionUID = 4341025008217142732L;
 
@@ -32,95 +30,15 @@ public class SubmissionStatusCodeEntity implements AbstractEntity {
 
 	@Id
 	@Column(name = "SUBMISSION_STATUS_CODE")
-	private String submissionStatusCode;
+	private String code;
 
-	@Column(name = "DESCRIPTION")
-	private String description;
-
-	@Column(name = "EFFECTIVE_DATE")
-	private Date effectiveDate;
-
-	@Column(name = "EXPIRY_DATE")
-	private Date expiryDate;
-
-	@Column(name = "CREATE_TIMESTAMP")
-	private Date createTimestamp;
-
-	@Column(name = "UPDATE_TIMESTAMP")
-	private Date updateTimestamp;
-
-	@Column(name = "CREATE_USER")
-	private String createUser;
-
-	@Column(name = "UPDATE_USER")
-	private String updateUser;
-
-	public String getSubmissionStatusCode() {
-		return submissionStatusCode;
+	@Override
+	public String getCode() {
+		return code;
 	}
 
-	public void setSubmissionStatusCode(String submissionStatusCode) {
-		this.submissionStatusCode = submissionStatusCode;
+	public void setCode(String code) {
+		this.code = code;
 	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	public Date getEffectiveDate() {
-		return effectiveDate;
-	}
-
-	public void setEffectiveDate(Date effectiveDate) {
-		this.effectiveDate = effectiveDate;
-	}
-
-	public Date getExpiryDate() {
-		return expiryDate;
-	}
-
-	public void setExpiryDate(Date expiryDate) {
-		this.expiryDate = expiryDate;
-	}
-
-	public Date getCreateTimestamp() {
-		return createTimestamp;
-	}
-
-	public void setCreateTimestamp(Date createTimestamp) {
-		this.createTimestamp = createTimestamp;
-	}
-
-	public Date getUpdateTimestamp() {
-		return updateTimestamp;
-	}
-
-	public void setUpdateTimestamp(Date updateTimestamp) {
-		this.updateTimestamp = updateTimestamp;
-	}
-
-	public String getCreateUser() {
-		return createUser;
-	}
-
-	public void setCreateUser(String createUser) {
-		this.createUser = createUser;
-	}
-
-	public String getUpdateUser() {
-		return updateUser;
-	}
-
-	public void setUpdateUser(String updateUser) {
-		this.updateUser = updateUser;
-	}
-
-
-	
-	
 
 }
