+++ backend/src/main/java/ca/bc/gov/app/m/postgres/client/service/ClientService.java
@@ -2,12 +2,12 @@ package ca.bc.gov.app.m.postgres.client.service;
 
 import java.util.List;
 
-import ca.bc.gov.app.m.postgres.client.entity.ClientTypeCodeEntity;
+import ca.bc.gov.app.core.vo.CodeDescrVO;
 
 public interface ClientService {
 	
 	String BEAN_NAME = "clientService";
 
-	List<ClientTypeCodeEntity> findActiveClientTypeCodes();
+	List<CodeDescrVO> findActiveClientTypeCodes();
 
 }
