+++ backend/src/main/java/ca/bc/gov/app/m/postgres/client/service/impl/ClientServiceImpl.java
@@ -8,6 +8,7 @@ import javax.inject.Inject;
 import org.springframework.stereotype.Service;
 
 import ca.bc.gov.app.core.util.CoreUtil;
+import ca.bc.gov.app.core.vo.CodeDescrVO;
 import ca.bc.gov.app.m.postgres.client.entity.ClientTypeCodeEntity;
 import ca.bc.gov.app.m.postgres.client.repository.ClientTypeCodeRepository;
 import ca.bc.gov.app.m.postgres.client.service.ClientService;
@@ -22,9 +23,10 @@ public class ClientServiceImpl implements ClientService {
 	private ClientTypeCodeRepository clientTypeCodeRepository;
 
 	@Override
-	public List<ClientTypeCodeEntity> findActiveClientTypeCodes() {
+	public List<CodeDescrVO> findActiveClientTypeCodes() {
 		Date currentTime = coreUtil.getCurrentTime();
-		return clientTypeCodeRepository.findActiveAt(currentTime);
+		List<ClientTypeCodeEntity> clientTypeCodes = clientTypeCodeRepository.findActiveAt(currentTime);
+		return coreUtil.toSortedCodeDescrVOs(clientTypeCodes);	
 	}
 
 }
